#4 ➕ ADD : Book & Author 엔티티

+++ src/main/java/com/brunch/server/author/entity/Author.java
@@ -13,5 +13,5 @@ public class Author {
     private String name;
     private String description;
     private String job;
-    private Long subscriber;
+    private int subscriber;
 }
